refactor(kernel-tui): split session-detail-view into focused files
The session-detail-view component file was 1235 lines, mixing the
top-level component, three subcomponents, pure formatting helpers,
layout math, and provision derivation. Split into a directory:

  session-detail-view/
    index.tsx               (barrel)
    session-detail-view.tsx (top-level component)
    format.ts               (pure formatting helpers)
    layout.ts               (entryRowCount, windowEndIdx, clampScroll)
    provisions.ts           (buildProvisions, provisionKey, etc.)
    provision-editor.tsx
    provisions-panel.tsx
    revoke-confirm.tsx

Layout-math and buildProvisions switched to options-bag signatures
along the way (CLAUDE.md compliance past two args). Adds 62 vitest
cases for the pure helpers (parseDescription, splitShellCommand,
formatExpandedContent, layout math, provision derivation). Adds
@ocap/repo-tools/test-utils/mock-endoify to the package's vitest
setupFiles so tests can import from @metamask/kernel-utils/session,
matching kernel-utils's own setup.
