+++ packages/kernel-tui/src/components/session-detail-view/index.tsx
@@ -1,0 +1,2 @@
+export { formatExpandedContent, parseDescription } from './format.ts';
+export { SessionDetailView } from './session-detail-view.tsx';
